Resolver moved to core
