+++ modules/libSceNet/CMakeLists.txt
@@ -6,10 +6,10 @@ project(${libName})
 
 add_library(${libName} SHARED entry.cpp wepoll.c epoll.cpp socket.cpp resolver.cpp)
 
-add_dependencies(${libName} third_party config_emu)
+add_dependencies(${libName} third_party config_emu core)
 
 target_compile_definitions(${libName} PUBLIC BOOST_ALL_NO_LIB WIN32_LEAN_AND_MEAN)
 
-target_link_libraries(${libName} ws2_32 libboost_thread config_emu.lib)
+target_link_libraries(${libName} ws2_32 libboost_thread config_emu.lib core.lib)
 
 setupModule(${libName})
